pyproject.toml
[tool.poetry]
name = "kwenta"
version = "0.1.0"
description = ""
authors = ["Kenneth V. Domingo <hello@kvdomingo.xyz>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.3"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.5"
python-dotenv = "^0.21.0"
django-cors-headers = "^3.13.0"
whitenoise = "^6.2.0"
gunicorn = {extras = ["gevent"], version = "^20.1.0"}
dj-database-url = "^1.0.0"
django-filter = "^22.1"

[tool.poetry.group.dev.dependencies]
black = {extras = ["d"], version = "^22.10.0"}
isort = "^5.10.1"
pre-commit = "^2.20.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ["py310"]

[tool.isort]
profile = "black"
